refactor magic_cmd.py
