+++ src/sql/cmd/cmd_utils.py
@@ -1,0 +1,12 @@
+import argparse
+import sys
+from sql import exceptions
+
+
+class CmdParser(argparse.ArgumentParser):
+    def exit(self, status=0, message=None):
+        if message:
+            self._print_message(message, sys.stderr)
+
+    def error(self, message):
+        raise exceptions.UsageError(message)
